防止确实mapfile导致的程序崩溃

+++ src/quene.js
@@ -104,39 +104,42 @@ function checkAndFillQuene() {
     if (count < 1000) {
       let counter = 0
       let mapFile = siteMapFiles.pop()
-      let filePath = path.resolve(siteMapDir, mapFile)
-      console.log(`开始从${mapFile}中获取url`)
-      
-      fs.readFile(filePath, (err, gzContent) => {
-        zlib.gunzip(gzContent, function (err, plainContent) {
-          console.log(`已解压文件，正在处理...`)
-          try {
-            let $ = cheerio.load(plainContent)
-            console.log(`文件解析完毕，开始判断链接...`)
-            $('loc').each(function (i, e) {
-              let $e = $(e)
-              let loc = $e.text()
-              let priority = parseFloat($(e).siblings('priority').text())
-              if (isValidUrl(loc)) {
-                counter++
-                quene.push({loc, priority})
-              }
-            })
-          }catch(e) {
-            console.log('加载文件出现错误, filename: ' + filePath)
-          }
-          
-          
-          console.log(`共添加了${counter}个记录`)
-          
-          fs.unlinkSync(filePath)
-          if (counter < 100) {
-            setTimeout(checkAndFillQuene, 1000)
-          }else {
-            setTimeout(checkAndFillQuene, 5000)
-          }
+      if (mapFile) {
+        let filePath = path.resolve(siteMapDir, mapFile)
+        console.log(`开始从${mapFile}中获取url`)
+        
+        fs.readFile(filePath, (err, gzContent) => {
+          zlib.gunzip(gzContent, function (err, plainContent) {
+            console.log(`已解压文件，正在处理...`)
+            try {
+              let $ = cheerio.load(plainContent)
+              console.log(`文件解析完毕，开始判断链接...`)
+              $('loc').each(function (i, e) {
+                let $e = $(e)
+                let loc = $e.text()
+                let priority = parseFloat($(e).siblings('priority').text())
+                if (isValidUrl(loc)) {
+                  counter++
+                  quene.push({loc, priority})
+                }
+              })
+            }catch(e) {
+              console.log('加载文件出现错误, filename: ' + filePath)
+            }
+            
+            
+            console.log(`共添加了${counter}个记录`)
+            
+            fs.unlinkSync(filePath)
+            if (counter < 100) {
+              setTimeout(checkAndFillQuene, 1000)
+            }else {
+              setTimeout(checkAndFillQuene, 5000)
+            }
+          })
         })
-      })
+      }
+      
     } else {
       setTimeout(checkAndFillQuene, 5000)
     }
